refactor(builder): share validity predicate machinery
Amp-Thread-ID: https://ampcode.com/threads/T-01a04442-3237-723e-8c8a-eafdc0f8743d
