+++ crates/builder/core/src/flashblocks/payload.rs
@@ -55,10 +55,8 @@ use tracing::{debug, error, info, metadata::Level, span, warn};
 use crate::{
     BuilderConfig, BuilderMetrics, ExecutionInfo, PayloadBuilder, ResourceLimits,
     flashblocks::{
-        FlashblocksExtraCtx,
-        best_txs::{BestFlashblocksTxs, ParkableBestPayloadTransactions},
-        context::BasePayloadBuilderCtx,
-        generator::BuildArguments,
+        BasePayloadBuilderCtx, BestFlashblocksTxs, FlashblocksExtraCtx,
+        ParkableBestPayloadTransactions, generator::BuildArguments,
     },
     traits::{ClientBounds, PoolBounds},
     transaction_events::{
