+++ crates/execution/payload/src/lib.rs
@@ -15,9 +15,24 @@ pub mod config;
 pub mod error;
 pub mod payload;
 pub use payload::{BaseBuiltPayload, BasePayloadBuilderAttributes};
+
+mod parkable;
+pub use parkable::{
+    ParkableBestPayloadTransactions, ParkablePayloadTransactions, PayloadTransactionInvalidated,
+};
+
+mod predicate_loads;
+pub use predicate_loads::{PredicateLoadTracker, PredicateReadRecorder};
+
 mod traits;
 pub use traits::*;
 mod types;
 pub use types::BasePayloadTypes;
+
+mod validity;
+pub use validity::{
+    ParkedPredicateIndex, StateChangeEffects, ValidityPredicateEvaluation, ValidityPredicateKey,
+};
+
 pub mod validator;
 pub use validator::BaseExecutionPayloadValidator;
